requirements.txt
discord.py
requests
asyncio
feedparser
moment